// src/generated/client/src/models/Customize.ts
// tslint:disable
/**
 * localhost
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Customize
 */
export interface Customize {
    /**
     * Showcase image url.
     * @type {string}
     * @memberof Customize
     */
    showcase_image?: string;
    /**
     * Showcase title text.
     * @type {string}
     * @memberof Customize
     */
    showcase_title?: string;
    /**
     * Showcase text.
     * @type {string}
     * @memberof Customize
     */
    showcase_text?: string;
    /**
     * Showcase button text.
     * @type {string}
     * @memberof Customize
     */
    showcase_button_text?: string;
    /**
     * Showcase button link.
     * @type {string}
     * @memberof Customize
     */
    showcase_button_link?: string;
}

export function CustomizeFromJSON(json: any): Customize {
    return CustomizeFromJSONTyped(json, false);
}

export function CustomizeFromJSONTyped(json: any, ignoreDiscriminator: boolean): Customize {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'showcase_image': !exists(json, 'showcase_image') ? undefined : json['showcase_image'],
        'showcase_title': !exists(json, 'showcase_title') ? undefined : json['showcase_title'],
        'showcase_text': !exists(json, 'showcase_text') ? undefined : json['showcase_text'],
        'showcase_button_text': !exists(json, 'showcase_button_text') ? undefined : json['showcase_button_text'],
        'showcase_button_link': !exists(json, 'showcase_button_link') ? undefined : json['showcase_button_link'],
    };
}

export function CustomizeToJSON(value?: Customize | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'showcase_image': value.showcase_image,
        'showcase_title': value.showcase_title,
        'showcase_text': value.showcase_text,
        'showcase_button_text': value.showcase_button_text,
        'showcase_button_link': value.showcase_button_link,
    };
}