// src/generated/client/src/models/Page.ts
export interface Page {
    /**
     * The date the object was published, in the site\'s timezone.
     * @type {Date}
     * @memberof Page
     */
    date?: Date;
    /**
     * The date the object was published, as GMT.
     * @type {Date}
     * @memberof Page
     */
    date_gmt?: Date;
    /**
     * 
     * @type {PostGuid}
     * @memberof Page
     */
    guid?: PostGuid;
    /**
     * Unique identifier for the object.
     * @type {number}
     * @memberof Page
     */
    id?: number;
    /**
     * URL to the object.
     * @type {string}
     * @memberof Page
     */
    link?: string;
    /**
     * The date the object was last modified, in the site\'s timezone.
     * @type {Date}
     * @memberof Page
     */
    modified?: Date;
    /**
     * The date the object was last modified, as GMT.
     * @type {Date}
     * @memberof Page
     */
    modified_gmt?: Date;
    /**
     * An alphanumeric identifier for the object unique to its type.
     * @type {string}
     * @memberof Page
     */
    slug?: string;
    /**
     * A named status for the object.
     * @type {string}
     * @memberof Page
     */
    status?: PageStatusEnum;
    /**
     * Type of Post for the object.
     * @type {string}
     * @memberof Page
     */
    type?: string;
    /**
     * A password to protect access to the content and excerpt.
     * @type {string}
     * @memberof Page
     */
    password?: string;
    /**
     * Permalink template for the object.
     * @type {string}
     * @memberof Page
     */
    permalink_template?: string;
    /**
     * Slug automatically generated from the object title.
     * @type {string}
     * @memberof Page
     */
    generated_slug?: string;
    /**
     * The ID for the parent of the object.
     * @type {number}
     * @memberof Page
     */
    parent?: number;
    /**
     * 
     * @type {PostTitle}
     * @memberof Page
     */
    title?: PostTitle;
    /**
     * 
     * @type {PostContent}
     * @memberof Page
     */
    content?: PostContent;
    /**
     * The ID for the author of the object.
     * @type {number}
     * @memberof Page
     */
    author?: number;
    /**
     * 
     * @type {PostExcerpt}
     * @memberof Page
     */
    excerpt?: PostExcerpt;
    /**
     * The ID of the featured media for the object.
     * @type {number}
     * @memberof Page
     */
    featured_media?: number;
    /**
     * Whether or not comments are open on the object.
     * @type {string}
     * @memberof Page
     */
    comment_status?: PageCommentStatusEnum;
    /**
     * Whether or not the object can be pinged.
     * @type {string}
     * @memberof Page
     */
    ping_status?: PagePingStatusEnum;
    /**
     * The order of the object in relation to other object of its type.
     * @type {number}
     * @memberof Page
     */
    menu_order?: number;
    /**
     * Meta fields.
     * @type {object}
     * @memberof Page
     */
    meta?: object;
    /**
     * The theme file to use to display the object.
     * @type {string}
     * @memberof Page
     */
    template?: string;
}
// ...
export enum PageStatusEnum {
    Publish = 'publish',
    Future = 'future',
    Draft = 'draft',
    Pending = 'pending',
    Private = 'private'
}
/**
* @export
* @enum {string}
*/
export enum PageCommentStatusEnum {
    Open = 'open',
    Closed = 'closed'
}
/**
* @export
* @enum {string}
*/
export enum PagePingStatusEnum {
    Open = 'open',
    Closed = 'closed'
}